Moved depecreated function into a separated folder, to be automatically called if needed
